Extract `printTable`
